src/clam_prototype/requirements.txt
click~=8.1
python-magic~=0.4
humanize~=4.7